feat/mux-player-react: cleanup styles. start moving things to shared for forward-looking goals.
