+++ packages/mux-player-react/src/components/Styles.tsx
@@ -1,12 +1,6 @@
 import React from "react";
 // @ts-ignore
-import cssStr from "../styles2.css";
-// import cssStr from "../styles.css";
-
-// declare module "*.css" {
-//     const content: string;
-//     export default content;
-// }
+import cssStr from "../styles.css";
 
 const Styles = () => {
   return <style>{cssStr}</style>;
